Add Figgle source generator
This adds a new NuGet package `Figgle.Generator` which allows Figgle text to be rendered during compilation time. This avoids the need to ship Figgle binaries with an application, so long as the strings being rendered are static. If strings are dynamic, you will still need the current `Figgle` package.