Remove HashCode package
I was seeing an error about the assembly being unavailable when running the generator. While I love the `HashCode` class, we only use it in one place so can avoid the dependency easily enough.

Also, use an explicit comparer in equality, as the default takes the user's current locale into account which can lead to inconsistent behaviour across machines.