Generate property accessors in FiggleFonts class
Uses a T4 template to generate property accessors for bundled fonts. Reduces the chance that this gets out of sync.

Because several file names are not valid C# identifiers, or because they're really composite names, we add `Aliases.csv` as a mechanism to override the default name that would otherwise be calculated.