Add sample projects and restructure solution
Creates a sample project for each of the primary modes of operating with Figgle:

- Dynamic text generation (the easy path)
- Static text generation
- Embedding specific fonts from the `Figgle.Fonts` package
- Embedding specific fonts from files on disk

Each sample project should work in isolation outside of this repo. We don't want our `Directory.Build.props` or `.editorconfig` files to apply to these projects. For that reason, we move source and test projcets into new `src/` and `tests/` folders.

Also convert the solution from `.sln` to `.slnx`. Requires VS 17.14.